Fixes to GUFieldPropagatorPool.  Refinements and cosmetics elsewhere.
- GUFieldPropagatorPool:
     fixes to Initialise() and Extend()
     changed variables (that compare to vector size) from unsigned to size_t

- GUFieldPropagator: new methods to access driver, epsilon

- Steppers: cleanup
     changed iterator to 'unsigned int'

- Tests: cleanup of unused variables
     test GU Propagate Field:
          * added option to check Pool - create 'copy' propagator and use it

     test Integration Driver: print 'goodAdvance'

- TMagErrorStepper
     Reduced verbosity of constructor

+++ magneticfield/inc/GUFieldPropagatorPool.h
@@ -31,7 +31,7 @@ class GUFieldPropagatorPool
     bool Initialize(unsigned int numThreads); 
      // Create new propagators for each thread !
     
-    void CheckIndex(int num){
+    void CheckIndex(size_t num){
        assert(num>=0);
        assert(num< fFieldPropagatorVec.size());
     }
@@ -46,7 +46,7 @@ class GUFieldPropagatorPool
     GUFieldPropagatorPool( GUFieldPropagator* prototype = 0); // , void** banks=0 );  // Ensure one per thread
     ~GUFieldPropagatorPool() {} 
 
-    void Extend(unsigned int Num);
+    void Extend(size_t Num);
      // Create additional propagators, so that total is 'Num'
   private:
     unsigned int fNumberPropagators;
